LogFiles class added
Here is the new logfile system, I replaced the old « LogFile » class by
the new one called « LogFiles ».
For some reason, I couldn’t rename it to « LogFile » without the « s ».
I also made small changes related to the way things are printed in the
console.
It works fine for 1000 iterations, and it’s even a bit faster (30 secs
instead of 53 secs).