Updated Version with Reinforcement Policy working
When generating a Table using Random Moves and Maps, we get a nearly
optimal algorithm.

The only problem is in the top right corner, when the Agent is on the
Left, somehow the Predator will stay in the corner instead of following
the Agent.

PS : Rules with Q-Learning doesn’t work, I didn’t test Rules Random &
Rules Q-Learning yet.